add combineReducer
