+++ bloco_16/day_5/content/dogs-app-redux/src/redux/reducers/index.js
@@ -1,20 +1,4 @@
-import { GET_IMAGE, REQUEST_IMAGE, FAILED_REQUEST } from '../actions';
+import { combineReducers } from 'redux';
+import { reducer } from './reducer';
 
-const initialState = {
-  isFetching: false,
-  imagePath: '',
-  error: '',
-};
-
-export function reducer(state = initialState, action) {
-  switch (action.type) {
-    case REQUEST_IMAGE:
-      return { ...state, isFetching: true };
-    case GET_IMAGE:
-      return { ...state, imagePath: action.payload, isFetching: false };
-    case FAILED_REQUEST:
-      return { ...state, error: action.payload, isFetching: false };
-    default:
-      return state;
-  }
-}
+export const rootReducer = combineReducers({ reducer });
